requirements.txt
flask==2.0.1
frame-sdk==0.1.0
openai-whisper==20231117
google-cloud-translate==3.12.0
python-dotenv==1.0.0